+++ .changeset/sixty-turtles-taste.md
@@ -1,0 +1,5 @@
+---
+"@boostxyz/sdk": minor
+---
+
+refactor utils, put interfaces and encoding functions with their respective consuming classes
