+++ packages/sdk/src/Actions/ERC721MintAction.ts
@@ -8,24 +8,37 @@ import {
   writeErc721MintActionValidate,
 } from '@boostxyz/evm';
 import { bytecode } from '@boostxyz/evm/artifacts/contracts/actions/ERC721MintAction.sol/ERC721MintAction.json';
-import type { Address, ContractEventName, Hex } from 'viem';
+import {
+  type Address,
+  type ContractEventName,
+  type Hex,
+  encodeAbiParameters,
+  toHex,
+} from 'viem';
 import type {
   DeployableOptions,
   GenericDeployableParams,
 } from '../Deployable/Deployable';
 import {
-  type ERC721MintActionPayload,
   type GenericLog,
   type ReadParams,
   RegistryType,
   type WriteParams,
-  prepareERC721MintActionPayload,
-  prepareERC721MintActionValidate,
 } from '../utils';
-import { ContractAction } from './ContractAction';
+import {
+  ContractAction,
+  type ContractActionPayload,
+  prepareContractActionPayload,
+} from './ContractAction';
 
-export { erc721MintActionAbi, prepareERC721MintActionPayload };
-export type { ERC721MintActionPayload };
+export { erc721MintActionAbi };
+/**
+ * `ERC721MintActionPayload` is a re-exported `ContractActionPayload`
+ *
+ * @export
+ * @typedef {ERC721MintActionPayload}
+ */
+export type ERC721MintActionPayload = ContractActionPayload;
 
 /**
  * A generic `viem.Log` event with support for `ERC721MintAction` event types.
@@ -166,16 +179,16 @@ export class ERC721MintAction extends ContractAction<
   }
 
   /**
-   * Validate that the action has been completed successfully
+   * Validate that the action has been completed successfully. This API is protected to prevent accidental signature burning.
    *
-   * @public
+   * @protected
    * @async
    * @param {Address} holder - The holder
    * @param {BigInt} tokenId - The token ID
    * @param {?WriteParams<typeof erc721MintActionAbi, 'validate'>} [params]
    * @returns {Promise<{ hash: `0x${string}`; result: boolean; }>} - True if the action has been validated for the user
    */
-  public async validate(
+  protected async validate(
     holder: Address,
     tokenId: bigint,
     params?: WriteParams<typeof erc721MintActionAbi, 'validate'>,
@@ -186,14 +199,14 @@ export class ERC721MintAction extends ContractAction<
   /**
    * Validate that the action has been completed successfully
    *
-   * @public
+   * @protected
    * @async
    * @param {Address} holder - The holder
    * @param {BigInt} tokenId - The token ID
    * @param {?WriteParams<typeof erc721MintActionAbi, 'validate'>} [params]
    * @returns {Promise<{ hash: `0x${string}`; result: boolean; }>} - True if the action has been validated for the user
    */
-  public async validateRaw(
+  protected async validateRaw(
     holder: Address,
     tokenId: bigint,
     params?: WriteParams<typeof erc721MintActionAbi, 'validate'>,
@@ -235,4 +248,45 @@ export class ERC721MintAction extends ContractAction<
       ...this.optionallyAttachAccount(options.account),
     };
   }
+}
+
+/**
+ * Encodes a payload to validate that an action has been completed successfully.
+ *
+ *
+ * @export
+ * @param {Address} holder - The holder address
+ * @param {bigint} payload - The token ID
+ * @returns {Hex} - The first 20 bytes of the payload will be the holder address and the remaining bytes must be an encoded token ID (uint256)
+ */
+export function prepareERC721MintActionValidate(
+  holder: Address,
+  payload: bigint,
+) {
+  return encodeAbiParameters(
+    [
+      { type: 'address', name: 'holder' },
+      { type: 'bytes', name: 'payload' },
+    ],
+    [holder, toHex(payload)],
+  );
+}
+
+/**
+ * Given a {@link ContractActionPayload}, properly encode a `ContractAction.InitPayload` for use with {@link ERC721MintAction} initialization.
+ *
+ * @param {ContractActionPayload} param0
+ * @param {bigint} param0.chainId - The chain ID on which the target exists
+ * @param {Address} param0.target - The target contract address
+ * @param {Hex} param0.selector - The selector for the function to be called
+ * @param {bigint} param0.value - The native token value to send with the function call
+ * @returns {*}
+ */
+export function prepareERC721MintActionPayload({
+  chainId,
+  target,
+  selector,
+  value,
+}: ContractActionPayload) {
+  return prepareContractActionPayload({ chainId, target, selector, value });
 }
